Downloads/Air-Guitar-main/air-guitar-production/requirements.txt
pyserial==3.5
numpy==1.24.3
sounddevice==0.4.6
PyYAML==6.0
flask==3.0.0
python-rtmidi==0.4.9
scipy==1.10.1
soundfile==0.12.1
